+++ srunner/scenarios/control_loss.py
@@ -34,7 +34,7 @@ class ControlLoss(BasicScenario):
 
     category = "ControlLoss"
 
-    timeout = 60            # Timeout of scenario in seconds
+    timeout = 120            # Timeout of scenario in seconds
 
     # ego vehicle parameters
     _no_of_jitter_actions = 20
@@ -84,7 +84,7 @@ class ControlLoss(BasicScenario):
             self._ego_vehicle_max_throttle = min(noise + self._dynamic_mean_for_throttle, 1)
 
             # turn vehicle
-            turn = SteerVehicle(self.ego_vehicle, self._ego_vehicle_max_steer, self._ego_vehicle_max_throttle, name='Steering ' + str(i))
+            turn = AddNoiseToVehicle(self.ego_vehicle, self._ego_vehicle_max_steer, self._ego_vehicle_max_throttle, name='Jittering ' + str(i))
 
             jitter_action = py_trees.composites.Parallel("Jitter Actions with Timeouts",
                                                          policy=py_trees.common.ParallelPolicy.SUCCESS_ON_ALL)