+++ CHANGELOG.md
@@ -27,6 +27,7 @@
     - NoSignalJunctionCrossing: hero vehicle must cross a non-signalized intersection
     - VehicleTurningRight: hero vehicle must react to a cyclist or pedestrian crossing ahead after a right turn
     - VehicleTurningLeft: hero vehicle must react to a cyclist or pedestrian crossing ahead after a left turn
+    - ControlLoss: Hero vehicle must react to a control loss and regain its control
 * Added atomic behaviors using py_trees behavior trees library
     - InTriggerRegion: new behavior to check if an object is within a trigger region
     - InTriggerDistanceToVehicle: check if a vehicle is within certain distance with respect to a reference vehicle
@@ -41,3 +42,4 @@
     - StopVehicle: stop vehicle
     - WaitForTrafficLightState: wait for the traffic light to have a given state
     - SyncArrival: sync the arrival of two vehicles to a given target
+    - AddNoiseToVehicle: Add noise to steer as well as throttle of the vehicle